# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.JobDeleteResponse:
    properties:
      message:
        example: Job removed from monitoring successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  controllers.JobNotFoundErrorResponse:
    properties:
      message:
        example: Job not found
        type: string
      success:
        example: false
        type: boolean
    type: object
  controllers.JobNotificationRequest:
    properties:
      fileName:
        type: string
      md5Hash:
        type: string
      success:
        type: boolean
    required:
    - fileName
    - md5Hash
    type: object
...
  controllers.JobStatusData:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      dbmgt_id:
        example: 1
        type: integer
      id:
        example: job_12345
        type: string
      status:
        example: completed
        type: string
    type: object
  controllers.JobStatusErrorResponse:
    properties:
      message:
        example: Job ID is required
        type: string
      success:
        example: false
        type: boolean
    type: object
  controllers.JobStatusListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/controllers.JobStatusData'
        type: array
      message:
        example: All jobs status retrieved successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  controllers.JobStatusResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  controllers.JobStatusSingleResponse:
    properties:
      data:
        $ref: '#/definitions/controllers.JobStatusData'
      message:
        example: Job status retrieved successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: DBF Artifact API
  title: dbfartifactapi
  version: "1.0"
paths:
  /api/jobs/{job_id}:
    delete:
      consumes:
      - application/json
      description: Remove a job from the monitoring system (typically for completed
        or failed jobs)
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.JobDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.JobStatusErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.JobNotFoundErrorResponse'
      summary: Delete job from monitoring
      tags:
      - job-status
  /api/jobs/{job_id}/notify:
    post:
      consumes:
      - application/json
      description: Allow external services to report job completion with file data
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      - description: Job completion notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/controllers.JobNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.JobStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.JobStatusErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.JobNotFoundErrorResponse'
      summary: Notify job completion
      tags:
      - job-status
  /api/jobs/{job_id}/status:
    get:
      consumes:
      - application/json
      description: Get the current status of a background job
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.JobStatusSingleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.JobStatusErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.JobNotFoundErrorResponse'
      summary: Get job status by ID
      tags:
      - job-status
  /api/jobs/dbmgt/{dbmgt_id}/status:
    get:
      consumes:
      - application/json
      description: Get all jobs associated with a specific database management ID
      parameters:
      - description: Database Management ID
        in: path
        name: dbmgt_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.JobStatusListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.JobStatusErrorResponse'
      summary: Get jobs by database management ID
      tags:
      - job-status
  /api/jobs/status:
    get:
      consumes:
      - application/json
      description: Get the current status of all background jobs. Supports optional
        pagination via query parameters 'page' and 'page_size'
      parameters:
      - description: Page number (1-indexed, optional)
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (optional, default: 10)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.JobStatusListResponse'
      summary: Get all jobs status
      tags:
      - job-status
...
swagger: "2.0"